+++ pkg/api/query_range.go
@@ -7,46 +7,44 @@ package api
 import (
 	"context"
 	"fmt"
-	"github.com/prometheus/client_golang/prometheus"
-	pgMetrics "github.com/timescale/promscale/pkg/pgmodel/metrics"
 	"net/http"
 	"time"
 
 	"github.com/NYTimes/gziphandler"
 	"github.com/pkg/errors"
+
 	"github.com/timescale/promscale/pkg/log"
 	"github.com/timescale/promscale/pkg/promql"
 )
 
-func QueryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Queryable, metrics *Metrics) http.Handler {
-	hf := corsWrapper(conf, queryRange(conf, queryEngine, queryable, metrics))
+func QueryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Queryable, updateMetrics func(handler, code string, duration float64)) http.Handler {
+	hf := corsWrapper(conf, queryRange(conf, queryEngine, queryable, updateMetrics))
 	return gziphandler.GzipHandler(hf)
 }
 
-func queryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Queryable, metrics *Metrics) http.HandlerFunc {
+func queryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Queryable, updateMetrics func(handler, code string, duration float64)) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		statusCode := "400"
+		begin := time.Now()
+		defer func() {
+			updateMetrics("/api/v1/query_range", statusCode, time.Since(begin).Seconds())
+		}()
 		start, err := parseTime(r.FormValue("start"))
 		if err != nil {
 			log.Info("msg", "Query bad request:"+err.Error())
 			respondError(w, http.StatusBadRequest, err, "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "400"}).Inc()
-			metrics.InvalidQueryReqs.Add(1)
 			return
 		}
 		end, err := parseTime(r.FormValue("end"))
 		if err != nil {
 			log.Info("msg", "Query bad request:"+err.Error())
 			respondError(w, http.StatusBadRequest, err, "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "400"}).Inc()
-			metrics.InvalidQueryReqs.Add(1)
 			return
 		}
 		if end.Before(start) {
 			err := errors.New("end timestamp must not be before start time")
 			log.Info("msg", "Query bad request:"+err.Error())
 			respondError(w, http.StatusBadRequest, err, "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "400"}).Inc()
-			metrics.InvalidQueryReqs.Add(1)
 			return
 		}
 
@@ -54,8 +52,6 @@ func queryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Query
 		if err != nil {
 			log.Info("msg", "Query bad request:"+err.Error())
 			respondError(w, http.StatusBadRequest, errors.Wrap(err, "param step"), "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "400"}).Inc()
-			metrics.InvalidQueryReqs.Add(1)
 			return
 		}
 
@@ -63,8 +59,6 @@ func queryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Query
 			err := errors.New("zero or negative query resolution step widths are not accepted. Try a positive integer")
 			log.Info("msg", "Query bad request:"+err.Error())
 			respondError(w, http.StatusBadRequest, err, "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "400"}).Inc()
-			metrics.InvalidQueryReqs.Add(1)
 			return
 		}
 
@@ -75,8 +69,6 @@ func queryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Query
 				"increasing the 'promql-max-points-per-ts' limit", conf.MaxPointsPerTs)
 			log.Info("msg", "Query bad request:"+err.Error())
 			respondError(w, http.StatusBadRequest, err, "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "400"}).Inc()
-			metrics.InvalidQueryReqs.Add(1)
 			return
 		}
 
@@ -87,8 +79,6 @@ func queryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Query
 			if err != nil {
 				log.Info("msg", "Query bad request"+err.Error())
 				respondError(w, http.StatusBadRequest, err, "bad_data")
-				pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "400"}).Inc()
-				metrics.InvalidQueryReqs.Add(1)
 				return
 			}
 
@@ -96,7 +86,6 @@ func queryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Query
 			defer cancel()
 		}
 
-		begin := time.Now()
 		qry, err := queryEngine.NewRangeQuery(
 			queryable,
 			r.FormValue("query"),
@@ -105,37 +94,35 @@ func queryRange(conf *Config, queryEngine *promql.Engine, queryable promql.Query
 			step,
 		)
 		if err != nil {
+			statusCode = "400"
 			log.Info("msg", "Query parse error: "+err.Error())
 			respondError(w, http.StatusBadRequest, err, "bad_data")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "400"}).Inc()
 			return
 		}
 
 		res := qry.Exec(ctx)
-		pgMetrics.QueryDuration.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range"}).Observe(time.Since(begin).Seconds())
 
 		if res.Err != nil {
 			log.Error("msg", res.Err, "endpoint", "query_range")
 			switch res.Err.(type) {
 			case promql.ErrQueryCanceled:
+				statusCode = "503"
 				respondError(w, http.StatusServiceUnavailable, res.Err, "canceled")
-				pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "503"}).Inc()
 				return
 			case promql.ErrQueryTimeout:
-				pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "503"}).Inc()
+				statusCode = "503"
 				respondError(w, http.StatusServiceUnavailable, res.Err, "timeout")
 				return
 			case promql.ErrStorage:
-				pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "500"}).Inc()
+				statusCode = "500"
 				respondError(w, http.StatusInternalServerError, res.Err, "internal")
 				return
 			}
+			statusCode = "422"
 			respondError(w, http.StatusUnprocessableEntity, res.Err, "execution")
-			pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "422"}).Inc()
 			return
 		}
-		pgMetrics.Query.With(prometheus.Labels{"type": "metric", "handler": "/api/v1/query_range", "code": "2xx"}).Inc()
-
+		statusCode = "2xx"
 		respondQuery(w, res, res.Warnings)
 	}
 }