+++ pkg/api/query_range_test.go
@@ -161,11 +161,8 @@ func TestRangedQuery(t *testing.T) {
 					Timeout:    timeout,
 				},
 			)
-			invalidQueryReqs := &mockMetric{}
-			metrics := &Metrics{
-				InvalidQueryReqs: invalidQueryReqs,
-			}
-			handler := queryRange(&Config{MaxPointsPerTs: 11000}, engine, query.NewQueryable(tc.querier, nil), metrics)
+
+			handler := queryRange(&Config{MaxPointsPerTs: 11000}, engine, query.NewQueryable(tc.querier, nil), mockUpdaterForQuery(&mockMetric{}, nil))
 			queryUrl := constructRangedQuery(tc.metric, tc.start, tc.end, tc.step, tc.timeout)
 			w := doRangedQuery(t, handler, queryUrl, tc.canceled)