+++ pkg/migrations/sql/versions/dev/0.8.1-dev/2-alter_promscale_instance_information_column.sql
@@ -1,0 +1,3 @@
+ALTER TABLE _ps_catalog.promscale_instance_information
+  RENAME COLUMN promscale_metrics_queries_executed_total
+  TO promscale_metrics_queries_success_total;